Move Gaphor main code to `gaphor.ui`
Now all startup code is closer together.
