+++ gaphor/__main__.py
@@ -1,4 +1,6 @@
-from gaphor import main
+import sys
+
+from gaphor.ui import main
 
 if __name__ == "__main__":
-    main()
+    main(sys.argv)
